src/requirements.txt
google-api-python-client==2.154.0
google-genai==1.24.0
python-dotenv==1.1.1
